requirements.txt
astroid==2.3.3
attrs==19.3.0
autopep8==1.5.1
boto3==1.12.30
botocore==1.15.36
certifi==2020.4.5.1
chardet==3.0.4
docutils==0.15.2
grpcio==1.27.2
-e git+git@github.com:Hydrospheredata/hydro-serving-integrations.git@0dedfbadf606ac3cd82b7ce68359bb1f48c800a7#egg=hydro_integrations
hydro-serving-grpc==2.1.0
idna==2.8
importlib-metadata==1.6.0
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
numpy==1.18.2
packaging==20.3
pandas==0.25.0
pluggy==0.13.1
protobuf==3.11.3
protobuf3-to-dict==0.1.5
py==1.8.1
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rope==0.16.0
s3transfer==0.3.3
sagemaker==1.52.1
scipy==1.4.1
six==1.14.0
smdebug-rulesconfig==0.1.2
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.9
wrapt==1.11.2
zipp==3.1.0
